:truck: reorganize action modules for clarity. 	remove `vsl::app_dir` action since `toml::app.dirpath` is the same.
