+++ vsmtp-rule-engine/src/tests/email/mod.rs
@@ -30,7 +30,7 @@ fn test_email_context() {
         &Some(rules_path!["main.vsl"]),
     )
     .unwrap();
-    let mut state = get_default_state("./tmp/app");
+    let (mut state, _) = get_default_state("./tmp/app");
 
     assert_eq!(re.run_when(&mut state, &StateSMTP::Connect), Status::Accept);
     state.get_context().write().unwrap().body = Body::Raw(String::default());
@@ -58,7 +58,7 @@ fn test_email_bcc() {
         &Some(rules_path!["bcc", "main.vsl"]),
     )
     .unwrap();
-    let mut state = get_default_state("./tmp/app");
+    let (mut state, _) = get_default_state("./tmp/app");
 
     assert_eq!(re.run_when(&mut state, &StateSMTP::PostQ), Status::Accept);
 }
@@ -70,10 +70,10 @@ fn test_email_add_get_set_header() {
         &Some(rules_path!["mutate_header", "main.vsl"]),
     )
     .unwrap();
-    let mut state = get_default_state("./tmp/app");
+    let (mut state, _) = get_default_state("./tmp/app");
 
     assert_eq!(re.run_when(&mut state, &StateSMTP::Connect), Status::Deny);
-    let mut state = get_default_state("./tmp/app");
+    let (mut state, _) = get_default_state("./tmp/app");
     state.get_context().write().unwrap().body = Body::Raw(String::default());
     assert_eq!(re.run_when(&mut state, &StateSMTP::PreQ), Status::Accept);
     state.get_context().write().unwrap().body = Body::Parsed(Box::new(Mail {
