+++ vsmtp-rule-engine/src/tests/mod.rs
@@ -49,7 +49,10 @@ pub mod helpers {
 
     use crate::rule_engine::RuleState;
 
-    pub(super) fn get_default_state(dirpath: impl Into<std::path::PathBuf>) -> RuleState<'static> {
+    /// create a rule engine state with it's associated configuration.
+    pub(super) fn get_default_state(
+        dirpath: impl Into<std::path::PathBuf>,
+    ) -> (RuleState<'static>, Config) {
         let config = Config::builder()
             .with_version_str("<1.0.0")
             .unwrap()
@@ -73,6 +76,6 @@ pub mod helpers {
             .validate()
             .unwrap();
 
-        RuleState::new(&config)
+        (RuleState::new(&config), config)
     }
 }
